add salt to config, rename bcrypt, some clean up
