+++ src/controllers/authController.js
@@ -1,10 +1,11 @@
-const x = require("bcrypt");
+const bcrypt = require("bcrypt");
+const config = require("../config");
 const httpStatus = require("http-status");
 
 class AuthController {
-  constructor({ User }, { bcrypt = x }) {
+  constructor({ User }, { authBcrypt = bcrypt }) {
     this.User = User;
-    this.bcrypt = bcrypt;
+    this.bcrypt = authBcrypt;
   }
 
   async register(req, res) {
@@ -16,7 +17,7 @@ class AuthController {
       return res
         .status(httpStatus.CONFLICT)
         .json({ message: "A user already registered with this email." });
-    const salt = await this.bcrypt.genSalt(10);
+    const salt = await this.bcrypt.genSalt(config.authentication.salt);
     const hashedPassword = await this.bcrypt.hash(password, salt);
 
     user = await this.User.create({
